Split decision trees between cost and lateness
The decision tree has become a bit overwhelmingly huge - now that we
have multiple tasks, we should have a separate decision tree for each.

- Split `DecisionTree` into `CostDecisionTree` and
  `LatenessDecisionTree`
- Create new `DecisionTree` to hold common functionality between all
  decision trees
- Modify shared controller examples for steps to take `DecisionTree`
  class into account
