+++ app/controllers/steps/cost/determine_cost_controller.rb
@@ -1,4 +1,4 @@
-class Steps::Cost::DetermineCostController < StepController
+class Steps::Cost::DetermineCostController < Steps::CostStepController
   def show
     raise 'No tribunal case in session' unless current_tribunal_case
