+++ app/controllers/steps/cost/must_challenge_hmrc_controller.rb
@@ -1,4 +1,4 @@
-class Steps::Cost::MustChallengeHmrcController < StepController
+class Steps::Cost::MustChallengeHmrcController < Steps::CostStepController
   def show
     raise 'No tribunal case in session' unless current_tribunal_case
   end
