+++ app/controllers/steps/cost/start_controller.rb
@@ -1,4 +1,4 @@
-class Steps::Cost::StartController < StepController
+class Steps::Cost::StartController < Steps::CostStepController
   def show
   end
 end
