+++ app/controllers/steps/cost_step_controller.rb
@@ -1,0 +1,7 @@
+class Steps::CostStepController < StepController
+  private
+
+  def decision_tree_class
+    CostDecisionTree
+  end
+end
