+++ app/controllers/steps/lateness_step_controller.rb
@@ -1,0 +1,7 @@
+class Steps::LatenessStepController < StepController
+  private
+
+  def decision_tree_class
+    LatenessDecisionTree
+  end
+end
