+++ spec/controllers/steps/cost/case_type_unchallenged_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Cost::CaseTypeUnchallengedController, type: :controller do
-    it_behaves_like 'an intermediate step controller', CaseTypeForm
+  it_behaves_like 'an intermediate step controller', CaseTypeForm, CostDecisionTree
 end
