+++ spec/controllers/steps/cost/challenged_decision_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Cost::ChallengedDecisionController, type: :controller do
-  it_behaves_like 'a start point step controller', ChallengedDecisionForm
+  it_behaves_like 'a start point step controller', ChallengedDecisionForm, CostDecisionTree
 end
