+++ spec/controllers/steps/cost/dispute_type_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Cost::DisputeTypeController, type: :controller do
-    it_behaves_like 'an intermediate step controller', DisputeTypeForm
+  it_behaves_like 'an intermediate step controller', DisputeTypeForm, CostDecisionTree
 end
