+++ spec/controllers/steps/cost/penalty_amount_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Cost::PenaltyAmountController, type: :controller do
-    it_behaves_like 'an intermediate step controller', PenaltyAmountForm
+  it_behaves_like 'an intermediate step controller', PenaltyAmountForm, CostDecisionTree
 end
