+++ spec/controllers/steps/lateness/in_time_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Lateness::InTimeController, type: :controller do
-  it_behaves_like 'an intermediate step controller', InTimeForm
+  it_behaves_like 'an intermediate step controller', InTimeForm, LatenessDecisionTree
 end
