+++ spec/controllers/steps/lateness/lateness_reason_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Lateness::LatenessReasonController, type: :controller do
-  it_behaves_like 'an intermediate step controller', LatenessReasonForm
+  it_behaves_like 'an intermediate step controller', LatenessReasonForm, LatenessDecisionTree
 end
